pyproject.toml
[project]
name = "moorcheh-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.0",
    "huggingface-hub>=0.36.0",
    "moorcheh-sdk>=1.2.2",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "pyjwt>=2.8.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "torch>=2.9.1",
    "transformers>=4.57.1",
]
